backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.65",
    "langgraph-cli>=0.1.66",
    "langgraph-sdk>=0.1.40",
    "pandas>=2.2.3",
    "jupyter>=1.1.1",
    "nbformat>=5.10.4",
    "langchain>=0.3.17",
    "langchain-openai>=0.2.13",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.6.0",
    "pymongo>=4.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.19",
    "authlib>=1.3.2",
    "httpx>=0.27.2",
    "websockets>=14.1",
    "pydantic-settings>=2.8.0",
    "email-validator>=2.2.0",
    "itsdangerous>=2.2.0",
    "pytest>=8.4.1",
    "mongomock-motor>=0.0.36",
    "dspy>=2.6.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "factory-boy>=3.3.0",
    "mongomock-motor>=0.0.35",
    "freezegun>=1.5.0",
]
